messaging_app copy/requirements.txt
Django>=5.0
djangorestframework
django-environ
drf-yasg
djangorestframework-simplejwt
faker
psycopg2-binary
